// src/glitchtip/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        EnvironmentProjectSchema: {
            /** Name */
            name: string;
            /** ID */
            id?: number | null;
            /**
             * Is Hidden
             * @default false
             */
            isHidden: boolean;
        };
// ...
        ProjectTeamSchema: {
            /** Name */
            name: string;
            /** Slug */
            slug?: string | null;
            /** Id */
            id: string;
            /**
             * Avatar
             * @default {
             *       "avatarType": "",
             *       "avatarUuid": null
             *     }
             */
            avatar: {
                [key: string]: string | null;
            };
            /**
             * Color
             * @default
             */
            color: string;
            /**
             * Features
             * @default []
             */
            features: string[];
            /**
             * Hasaccess
             * @default true
             */
            hasAccess: boolean;
            /**
             * Isbookmarked
             * @default false
             */
            isBookmarked: boolean;
            /**
             * Isinternal
             * @default false
             */
            isInternal: boolean;
            /** Ismember */
            isMember: boolean;
            /**
             * Ispublic
             * @default false
             */
            isPublic: boolean;
            /** Scrubipaddresses */
            scrubIPAddresses: boolean;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Platform */
            platform?: string | null;
            /** First Event */
            firstEvent?: string | null;
            /**
             * Event Throttle Rate
             * @description Probability (in percent) on how many events are throttled. Used for throttling at project level
             * @default 0
             */
            eventThrottleRate: number;
            /** Teams */
            teams: components["schemas"]["TeamSlugSchema"][];
        };
        /** TeamSchema */
        TeamSchema: {
            /** Id */
            id: string;
            /** Slug */
            slug: string;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Ismember */
            isMember: boolean;
            /** Membercount */
            memberCount: number;
        };
// ...
        ProjectOrganizationSchema: {
            /** Name */
            name: string;
            /** Slug */
            slug?: string | null;
            /** Id */
            id: string;
            /**
             * Avatar
             * @default {
             *       "avatarType": "",
             *       "avatarUuid": null
             *     }
             */
            avatar: {
                [key: string]: string | null;
            };
            /**
             * Color
             * @default
             */
            color: string;
            /**
             * Features
             * @default []
             */
            features: string[];
            /**
             * Hasaccess
             * @default true
             */
            hasAccess: boolean;
            /**
             * Isbookmarked
             * @default false
             */
            isBookmarked: boolean;
            /**
             * Isinternal
             * @default false
             */
            isInternal: boolean;
            /** Ismember */
            isMember: boolean;
            /**
             * Ispublic
             * @default false
             */
            isPublic: boolean;
            /** Scrubipaddresses */
            scrubIPAddresses: boolean;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Platform */
            platform?: string | null;
            /** First Event */
            firstEvent?: string | null;
            /**
             * Event Throttle Rate
             * @description Probability (in percent) on how many events are throttled. Used for throttling at project level
             * @default 0
             */
            eventThrottleRate: number;
            organization: components["schemas"]["OrganizationSchema"];
        };
// ...
        ProjectSchema: {
            /** Name */
            name: string;
            /** Slug */
            slug?: string | null;
            /** Id */
            id: string;
            /**
             * Avatar
             * @default {
             *       "avatarType": "",
             *       "avatarUuid": null
             *     }
             */
            avatar: {
                [key: string]: string | null;
            };
            /**
             * Color
             * @default
             */
            color: string;
            /**
             * Features
             * @default []
             */
            features: string[];
            /**
             * Hasaccess
             * @default true
             */
            hasAccess: boolean;
            /**
             * Isbookmarked
             * @default false
             */
            isBookmarked: boolean;
            /**
             * Isinternal
             * @default false
             */
            isInternal: boolean;
            /** Ismember */
            isMember: boolean;
            /**
             * Ispublic
             * @default false
             */
            isPublic: boolean;
            /** Scrubipaddresses */
            scrubIPAddresses: boolean;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /** Platform */
            platform?: string | null;
            /** First Event */
            firstEvent?: string | null;
            /**
             * Event Throttle Rate
             * @description Probability (in percent) on how many events are throttled. Used for throttling at project level
             * @default 0
             */
            eventThrottleRate: number;
        };
        /** KeyRateLimit */
        KeyRateLimit: {
            /** Window */
            window: number;
            /** Count */
            count: number;
        };
        /**
         * ProjectKeySchema
         * @description A project key (DSN) provides a public authentication string used for event
         *     ingestion.
         */
        ProjectKeySchema: {
            /** Name */
            name?: string | null;
            rateLimit?: components["schemas"]["KeyRateLimit"] | null;
            /**
             * Datecreated
             * Format: date-time
             */
            dateCreated: string;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Dsn */
            dsn: {
                [key: string]: string;
            };
            /** Label */
            label: string | null;
            /**
             * Public
             * Format: uuid
             */
            public: string;
            /** Projectid */
            projectID: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}